+++ src/components/manage/Form/ModalForm.test.jsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import ModalForm from './ModalForm';
+
+jest.mock('./Field', () => jest.fn(() => <div className="Field" />));
+
+describe('ModalForm', () => {
+  it('renders a modal form component', () => {
+    const component = renderer.create(
+      <ModalForm
+        schema={{
+          fieldsets: [
+            {
+              id: 'default',
+              title: 'Default',
+              fields: ['title'],
+            },
+          ],
+          properties: {
+            title: {},
+          },
+          required: [],
+        }}
+        onSubmit={() => {}}
+        onCancel={() => {}}
+        open={false}
+      />,
+    );
+    const json = component.toJSON();
+    expect(json).toMatchSnapshot();
+  });
+});